workflowdemo/requirements.txt
certifi==2018.10.15
chardet==3.0.4
#Django==2.1.2
idna==2.7
#mysqlclient==1.3.13
pytz==2018.5
requests==2.19.1
urllib3==1.23
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
Pillow==5.3.0
#uwsgi
